fixing bugs on crud produtos

+++ server/controllers/controllerProdutos.go
@@ -9,6 +9,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/mquelucci/projeto-loja-virtual/server/database"
 	"github.com/mquelucci/projeto-loja-virtual/server/models"
+	"github.com/mquelucci/projeto-loja-virtual/server/utils"
 )
 
 func BuscarProdutos() []models.Produto {
@@ -20,24 +21,32 @@ func BuscarProdutos() []models.Produto {
 func CriarProduto(c *gin.Context) {
 	var produto models.Produto
 	descricao := c.PostForm("descricao")
-	preco, _ := strconv.ParseFloat(c.PostForm("preco"), 64)
-	quantidade, _ := strconv.Atoi(c.PostForm("quantidade"))
+	err := utils.ProdutoDuplo(descricao, false, &produto)
+	if err != nil {
+		c.HTML(http.StatusBadRequest, "novosProdutos.html", gin.H{
+			"configs": BuscarConfigs(),
+			"erro":    err.Error(),
+		})
+		return
+	}
+	produto.Descricao = descricao
+
 	imagem, err := c.FormFile("imagem")
 	if err != nil {
-		log.Println("Nenhum arquivo carregado. Salvando produto no banco de dados.")
+		log.Println("Nenhum arquivo carregado. Criando produto sem imagem.")
 		produto.Imagem = "/assets/images/not_found.png"
 	} else {
-		savePath := "./client/assets/images/" + imagem.Filename
-		err = c.SaveUploadedFile(imagem, savePath)
+		err := utils.TratarImagemProduto(c, imagem, &produto)
 		if err != nil {
 			c.HTML(http.StatusBadRequest, "novosProdutos.html", gin.H{
 				"configs": BuscarConfigs(),
-				"erro":    "Erro ao salvar a imagem" + err.Error(),
+				"erro":    err.Error(),
 			})
 			return
 		}
 		produto.Imagem = "/assets/images/" + imagem.Filename
 	}
+	preco, _ := strconv.ParseFloat(c.PostForm("preco"), 64)
 
 	if preco == 0.0 {
 		c.HTML(http.StatusBadRequest, "novosProdutos.html", gin.H{
@@ -46,9 +55,9 @@ func CriarProduto(c *gin.Context) {
 		})
 		return
 	}
-
-	produto.Descricao = descricao
 	produto.Preco = preco
+
+	quantidade, _ := strconv.Atoi(c.PostForm("quantidade"))
 	produto.Quantidade = int(quantidade)
 
 	if err := models.ValidaProduto(&produto); err != nil {
@@ -58,12 +67,14 @@ func CriarProduto(c *gin.Context) {
 		})
 		return
 	}
+
 	err = database.DB.Create(&produto).Error
 	if err != nil {
 		c.HTML(http.StatusBadRequest, "novosProdutos.html", gin.H{
 			"configs": BuscarConfigs(),
 			"erro":    err.Error(),
 		})
+		return
 	}
 	c.HTML(http.StatusCreated, "novosProdutos.html", gin.H{
 		"configs": BuscarConfigs(),
@@ -76,30 +87,36 @@ func EditarProduto(c *gin.Context) {
 	var produto models.Produto
 	id := c.Query("id")
 	database.DB.First(&produto, id)
+
 	descricao := c.PostForm("descricao")
-	preco, _ := strconv.ParseFloat(c.PostForm("preco"), 64)
-	quantidade, _ := strconv.Atoi(c.PostForm("quantidade"))
+	err := utils.ProdutoDuplo(descricao, true, &produto)
+	if err != nil {
+		c.HTML(http.StatusBadRequest, "editarProduto.html", gin.H{
+			"configs": BuscarConfigs(),
+			"produto": produto,
+			"erro":    err.Error(),
+		})
+		return
+	}
+	produto.Descricao = descricao
 
 	imagem, err := c.FormFile("imagem")
 	if err != nil {
 		log.Println("Nenhum arquivo carregado. Mantendo o registro de imagem do produto.")
 	} else {
-
-		savePath := "./client/assets/images/" + imagem.Filename
-		if savePath == produto.Imagem {
-			os.Remove(savePath)
-		}
-		err = c.SaveUploadedFile(imagem, savePath)
+		err := utils.TratarImagemProduto(c, imagem, &produto)
 		if err != nil {
-			c.HTML(http.StatusBadRequest, "novosProdutos.html", gin.H{
+			c.HTML(http.StatusBadRequest, "editarProduto.html", gin.H{
 				"configs": BuscarConfigs(),
-				"erro":    "Erro ao salvar a imagem" + err.Error(),
+				"produto": produto,
+				"erro":    err.Error(),
 			})
 			return
 		}
 		produto.Imagem = "/assets/images/" + imagem.Filename
 	}
 
+	preco, _ := strconv.ParseFloat(c.PostForm("preco"), 64)
 	if preco == 0.0 {
 		c.HTML(http.StatusBadRequest, "editarProduto.html", gin.H{
 			"configs": BuscarConfigs(),
@@ -108,9 +125,9 @@ func EditarProduto(c *gin.Context) {
 		})
 		return
 	}
-
-	produto.Descricao = descricao
 	produto.Preco = preco
+
+	quantidade, _ := strconv.Atoi(c.PostForm("quantidade"))
 	produto.Quantidade = int(quantidade)
 
 	if err := models.ValidaProduto(&produto); err != nil {
@@ -129,6 +146,7 @@ func EditarProduto(c *gin.Context) {
 			"produto": produto,
 			"erro":    err.Error(),
 		})
+		return
 	}
 	c.HTML(http.StatusAccepted, "editarProduto.html", gin.H{
 		"configs": BuscarConfigs(),
@@ -141,31 +159,41 @@ func RemoverImagemProduto(c *gin.Context) {
 	id := c.Query("id")
 	var produto models.Produto
 	database.DB.First(&produto, id)
-	pathImagem := "./client" + produto.Imagem
-	os.Remove(pathImagem)
-	produto.Imagem = "/assets/images/not_found.png"
-
-	err := database.DB.Save(&produto).Error
-	if err != nil {
-		c.HTML(http.StatusBadRequest, "editarProduto.html", gin.H{
-			"configs": BuscarConfigs(),
-			"produto": produto,
-			"erro":    err.Error(),
-		})
+	if produto.Imagem != "/assets/images/not_found.png" {
+		pathImagem := "./client" + produto.Imagem
+		os.Remove(pathImagem)
+		produto.Imagem = "/assets/images/not_found.png"
+		err := database.DB.Save(&produto).Error
+		if err != nil {
+			c.HTML(http.StatusBadRequest, "editarProduto.html", gin.H{
+				"configs": BuscarConfigs(),
+				"produto": produto,
+				"erro":    err.Error(),
+			})
+			return
+		}
 	}
-
 	c.HTML(http.StatusAccepted, "editarProduto.html", gin.H{
 		"configs": BuscarConfigs(),
 		"produto": produto,
 		"message": "Imagem removida com sucesso",
 	})
+
 }
 
 func DeletarProduto(c *gin.Context) {
 	id := c.Query("id")
 	var produto models.Produto
-	database.DB.Delete(&produto, id)
-	c.HTML(http.StatusAccepted, "adminProdutos.html", gin.H{
+	err := database.DB.Delete(&produto, id).Error
+	if err != nil {
+		c.HTML(http.StatusBadRequest, "adminProdutos.html", gin.H{
+			"configs":  BuscarConfigs(),
+			"produtos": BuscarProdutos(),
+			"erro":     err.Error(),
+		})
+		return
+	}
+	c.HTML(http.StatusBadRequest, "adminProdutos.html", gin.H{
 		"configs":  BuscarConfigs(),
 		"produtos": BuscarProdutos(),
 	})
